Correct answer on ara standard example
At this point, the program is able to regenerate the ara standard
example gene-gene network and get the right answer, with the caveat that
any self-interactions of the genes are not included in the edge list.
Self loops are illegal in simple networks, so I will probably have to
build a new network type.